requirements.txt
black==24.8.0
click==8.1.7
loguru==0.7.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
